Update Office.FileDialogFilters.md
I've added the warning at the top as it is so easy to overlook this, and it should be stated up-front. This has been a bug for at least 10 years. It would be nice if the Save As finally started to work for .clear, .add, and .delete too. See e.g.:
https://stackoverflow.com/questions/45332735/vba-excel-filedialog-to-set-reset-filters