Update customize-the-office-fluent-ribbon-by-using-an-open-xml-formats-file.md
The code samples are for Office 2007. Surely in 2020 we would like people to use the 2010 and later XML schema?!
Also please resurrect the OfficeCustomUIEditorSetup.msi the links have gone (e.g. https://www.microsoft.com/en-us/download/details.aspx?id=50745) and it is no GitHub (https://github.com/OfficeDev/office-custom-ui-editor) but not in an executable format. It should be.